Use maximum color value for luminance
Instead of using a luminance approximation based on the color channels, just use the maximum value from the color channels as luminance.
Move the horizontal luminance calculation back into the horizontal upscale. The benefits of calculating it when we shift in new pixel data have vanished and the timing is more difficult there.