Use half size when height matches vertical resolution. (#67)
If the height exactly matches the vertical resolution then v_osd_start will be set to 0, but v_cnt starts counting from 1 so the OSD will never be drawn. Maybe v_cnt should start at 0, but that will throw of some other things. This fixes the issue I am seeing in the M72 core which has a vertical resolution of 256.