CMD24 (state WR_STATE_SEND_DRESP) fix (#86)
* courtesy of viandant
The response code from module sd_card sent at the end of executing
CMD24 (state WR_STATE_SEND_DRESP), was not a byte starting with bits 1110
as expected by the sd_controller module. This caused the sd_controller
to assume the result byte is finished and switch off the clock signal prematurely.
After that the system could not recover and return to a working state.

* from jotego

Incidentally, all these wire constants should be LOCALPARAM instead.  Changed.

---------

Co-authored-by: Fred VanEijk <fvaneijk.usa@gmail.com>